requirements.txt
flask
scikit-learn==1.2.2
numpy
matplotlib.pyplot
cv2
keras